Fix sense of locks
